src/main/java/com/crud/kodilla_library/mapper/RentalMapper.java
package com.crud.kodilla_library.mapper;

import com.crud.kodilla_library.domain.Rental;
import com.crud.kodilla_library.domain.dto.RentalDto;

import org.springframework.stereotype.Service;

import java.util.List;
import java.util.stream.Collectors;

@Service
public class RentalMapper {

    public Rental mapToRental(final RentalDto rentalDto) {
        return new Rental(
                rentalDto.getRentId(),
                rentalDto.getExemplar(),
                rentalDto.getUser(),
                rentalDto.getRentDate(),
                rentalDto.getReturnDate()
        );
    }

    public RentalDto mapToRentalDto(final Rental rental) {
        return new RentalDto(
                rental.getRentId(),
                rental.getExemplar(),
                rental.getUser(),
                rental.getRentDate(),
                rental.getReturnDate()
        );
    }

    public List<RentalDto> mapToRentalDtoList(final List<Rental> rentalList) {
        return rentalList.stream()
                .map(this::mapToRentalDto)
                .collect(Collectors.toList());
    }
}
